+++ DesignToolbox/DesignToolboxSnapshotsTests/__Snapshots__/Components/OrangeCompactTheme/Components/Controls/OrangeCompactTheme+PasswordInputSnapshotsTests.swift
@@ -1,0 +1,46 @@
+//
+// Software Name: OUDS iOS
+// SPDX-FileCopyrightText: Copyright (c) Orange SA
+// SPDX-License-Identifier: MIT
+//
+// This software is distributed under the MIT license,
+// the text of which is available at https://opensource.org/license/MIT/
+// or see the "LICENSE" file for more details.
+//
+// Authors: See CONTRIBUTORS.txt
+// Software description: A SwiftUI components library with code examples for Orange Unified Design System
+//
+
+import OUDSSwiftUI
+import SwiftUI
+
+// swiftlint:disable required_deinit
+// swiftlint:disable type_name
+// MARK: - Test Cases
+
+/// Tests the UI rendering of the `OUDSPasswordInput` for each parameter for `OrangeCompactTheme`.
+final class OrangeCompactThemePasswordInputSnapshotsTests: PasswordInputSnapshotsTestsTestCase {
+
+    // swiftlint:disable implicitly_unwrapped_optional
+    private var theme: OUDSTheme!
+    // swiftlint:enable implicitly_unwrapped_optional
+
+    override func setUp() {
+        theme = OrangeCompactTheme()
+    }
+
+    /// Tests all password inputs configurations in the `OrangeCompactTheme` with the `light` color schemes.
+    @MainActor func testAllPasswordInputsOrangeCompactThemeLight() {
+        let interfaceStyle = UIUserInterfaceStyle.light
+        testAllPasswordInputs(theme: theme, interfaceStyle: interfaceStyle)
+    }
+
+    /// Tests all password inputs configurations in the `OrangeCompactTheme` with the `dark` color schemes.
+    @MainActor func testAllPasswordInputsOrangeCompactThemeDark() {
+        let interfaceStyle = UIUserInterfaceStyle.dark
+        testAllPasswordInputs(theme: theme, interfaceStyle: interfaceStyle)
+    }
+}
+
+// swiftlint:enable required_deinit
+// swiftlint:enable type_name