+++ CHANGELOG.md
@@ -4,6 +4,21 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.2.0](https://github.com/Orange-OpenSource/ouds-ios-design-system-toolbox/compare/1.1.0...1.2.0) - 2026-02-13
+
+### Added
+
+- [Library] `password input` component (Orange-OpenSource/ouds-ios#997)
+- [Library] `bullet list` component (Orange-OpenSource/ouds-ios#513)
+
+### Changed
+
+- [DesignToolbox] `ruby/setup-ruby` action from v1.286.0 to v1.288.0 for `build-and-test` workflow
+- [DesignToolbox] `github/codeql-action/upload-sarif` action for `scorecard` workflow
+- [DesignToolbox] `json` RubyGem from v2.18.0 to v2.18.1
+- [DesignToolbox] Update `fastlane` gem from v2.231.1 to v2.232.1
+- [DesignToolbox] Rename of *Orange Business Tools*  theme to *Orange Compact* (Orange-OpenSource/ouds-ios#1292)
+
 ## [1.1.0](https://github.com/Orange-OpenSource/ouds-ios-design-system-toolbox/compare/1.0.0...1.1.0) - 2026-01-27
 
 ### Added